requirements.txt
streamlit==1.36.0
google-cloud-vision
groq

